Add support for scope-aware variable stacks
JSHint is very useful at spotting accidentally shadowed or mistyped
variables. To port that functionality to Next we need to have a
scope-aware variable stacks: so that at any time we can easily check
if a variable has already been defined and is accesible from within
the current scope.

This patchs adds ScopeStack object to maintain such stacks.