New Tokens class for easier tokens traversal
It is a real pain in the ass to manually calculate indexes and
always keep local variables referencing previous, next and current
tokens. This patch adds a new object, Tokens, that is essentially
a wrapper around an array of tokens with little helper functions
such as peak(), next(), prev(), etc.