Improve safe() function to handle all default Object properties.
Jamund (gh:xjamnundx) found out that ScopeStack.addUse method fails
on variables such as toString, etc. This happens because we're to
add a new property with a name that is similar to one already in
a prototype chain for this object. We had this problem with __proto__
before and that's why we have function safe() that handles this
particular case.

This patch improves safe() to handle all default Object properties.

Issues: GH-6