Add new module: regexp.
So far it is pretty primitive--I reimplemented option 'regexp' but
hopefully in future it will be able to catch more subtle regular
expression oddities.

Also Esprima blows up on invalid regular expression so we don't need
to parse/verify it ourselves. However, this means that we really have
to make Esprima more tolerant.