Replace Events library with EventEmitter.
There is no need to include a library we extracted from Backbone
if Node has a perfectly fine built-in helper called EventEmitter.

Well, perfectly fine is a bit of a stretch because it doesn't
support space separated events:

  linter.on("IfStatetement ForStatement", function () { ... });

So when I was swaping helpers I had to write wrapper functions for
the Linter object: on and trigger that act as their counterparts in
Backbone.

I also used composition instead of inheritance.