Switch to the bleeding edge version of Esprima.
We need a few Esprima features that are not in the stable version
yet so I've decided to use -dev version until Esprima 1.0 arrives.

Esprima changed how its output data a bit so I need to update some
code to make sure all tests pass (they do now). But this means
that JSHint Next is not compatible with current stable version of
Esprima anymore.